+++ GPX/Helpers/GPXOptimize.cs
